move software/painter tests to test package
